+++ R/bar_plot.R
@@ -1,7 +1,8 @@
+#' @importFrom ggplot2 ylim
 # bar plot function
 bar_plot <- function(df_plot, x, y, fill, label, xlb = "", ylb = "", 
                      ttl = "", sttl = "", lgnd = NULL, rotate = FALSE, 
-                     col_palette = NULL){
+                     col_palette = NULL, ylim_range = NULL){
   # basic bar plot
   plt <- df_plot %>% 
     ggplot(aes_string(x = x, y = y, fill = fill, label = label)) + 
@@ -14,6 +15,10 @@ bar_plot <- function(df_plot, x, y, fill, label, xlb = "", ylb = "",
     plt <- plt + guides(fill = FALSE)
   } else {
     plt <- plt + scale_fill_discrete(name = lgnd)
+  }
+  # apply lower limit to the y-axis?
+  if(!is.null(ylim_range)){
+    plt <- plt + ylim(ylim_range)
   }
   # rotate x-axis labels if requested
   if(rotate){